Moved Specification interface to Specification folder
